Add ParseCondExpr method to parser
It seems to be working?? maybe?? i didnt really expect it to work.
Modified ParseExpr to account for Keywords in Expr as well.

WARNING: Lexer::Tokenize breaks down if the input string is not
terminated by a semicolon or RPAREN
This is sth to look out for when using it for testing,
but i think we can maybe sidestep this issue? idk

feels like it might take a lot of time to deal with